requirements.txt
awsebcli==3.3.2
cement==2.4.0
docker-py==1.1.0
dockerpty==0.3.2
docopt==0.6.2
facebook-sdk==0.4.0
fitbit==0.1.0
Flask==0.10.1
Flask-Login==0.2.9
Flask-Stormpath==0.3.6
Flask-WTF==0.11
httplib2==0.9
itsdangerous==0.24
Jinja2==2.7.3
jmespath==0.7.1
MarkupSafe==0.23
oauth2client==1.2
oauthlib==0.6.3
pathspec==0.3.3
phue==0.7
py-bcrypt==0.4
pyhue==0.7.6
PyJWT==0.4.1
PyMySQL==0.6.4
python-dateutil==2.4.2
PyYAML==3.11
requests==2.6.2
requests-oauthlib==0.4.1
six==1.9.0
stormpath==1.3.5
texttable==0.8.3
virtualenv==12.0.7
websocket-client==0.30.0
Werkzeug==0.10
WTForms==2.0.2
yummly==0.5.0
